Change the format from xml to json
Here, in the request to the Nominatim server, we specify the format as
JSON as is easier to extract data from this format.

Also, the zoom levels for this program vary from 0-18 (same as
Nominatim), as compared to the 1-5 zoom levels that were available for
the program using XML format. This reduces a number of issues we had
with parsing the data when it was in XML format.